+++ substratevm/CHANGELOG.md
@@ -4,6 +4,8 @@ This changelog summarizes major changes to GraalVM Native Image.
 
 ## Version 23.1.0
 * (GR-35746) Lower the default aligned chunk size from 1 MB to 512 KB for the serial and epsilon GCs, reducing memory usage and image size in many cases.
+* (GR-45014) Support JFR on Windows.
+* (GR-45015) Support the heap dumping infrastructure on Windows.
 
 ## Version 23.0.0
 * (GR-40187) Report invalid use of SVM specific classes on image class- or module-path as error. As a temporary workaround, `-H:+AllowDeprecatedBuilderClassesOnImageClasspath` allows turning the error into a warning.